+++ src/Prato.java
@@ -2,10 +2,10 @@ import org.json.JSONObject;
 
 public class Prato implements JsonFormatter {
     private String nome;
-    private float preco;
+    private double preco;
 
     //Construtor com nome e preco
-    public Prato(String nome, float preco) throws ExceptionPrato {
+    public Prato(String nome, double preco) throws ExceptionPrato {
         setNome(nome);
         setPreco(preco);
     }
@@ -13,7 +13,7 @@ public class Prato implements JsonFormatter {
     //Construtor vazio
     public Prato() throws ExceptionPrato{
         setNome("Comida");
-        setPreco(0.1F);
+        setPreco(0.1);
     }
 
     //Alterar nome
@@ -25,7 +25,7 @@ public class Prato implements JsonFormatter {
     }
 
     //Alerar preco
-    public void setPreco(float preco) throws ExceptionPrato {
+    public void setPreco(double preco) throws ExceptionPrato {
         if (preco > 0)
             this.preco = preco;
         else
@@ -38,7 +38,7 @@ public class Prato implements JsonFormatter {
     }
 
     //Pegar preco
-    public float getPreco() {
+    public double getPreco() {
         return preco;
     }
